fix: auth layout
