+++ src/modules/auth/index.ts
@@ -1,3 +1,4 @@
-import { AuthFormContainer } from './ui/AuthFormContainer/AuthFormContainer';
+import { AuthContainer } from './ui/AuthContainer/AuthContainer';
+import { AuthLayout } from './ui/AuthLayout/AuthLayout';
 import { AuthProvider } from './ui/AuthProvider/AuthProvider';
-export { AuthProvider, AuthFormContainer };
+export { AuthProvider, AuthContainer, AuthLayout };
