+++ src/modules/auth/ui/AuthContainer/AuthContainer.module.css
@@ -1,0 +1,4 @@
+.AuthContainer {
+    display: flex;
+    background-color: var(--neutral-gray-1);
+}
