--- src/modules/auth/ui/AuthFormContainer/AuthForm.module.css
@@ -1,3 +1,0 @@
-.AuthForm {
-    display: flex;
-}
