+++ src/modules/auth/ui/AuthLayout/AuthLayout.module.css
@@ -1,0 +1,26 @@
+.overlay {
+    width: 100%;
+    height: 100%;
+    background: none;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    position: relative;
+    min-height: 100vh;
+}
+
+.bg {
+    filter: blur(8px);
+    -webkit-filter: blur(8px);
+    background-position: center;
+    background-repeat: no-repeat;
+    background-size: cover;
+    width: 100%;
+    height: 100%;
+    position: absolute;
+}
+
+.content {
+    padding: 0;
+    position: absolute;
+}
