+++ src/modules/auth/ui/AuthProvider/AuthProvider.tsx
@@ -1,8 +1,8 @@
 import { LS_AuthKey } from '@shared/constants/constants';
 import { useAppDispatch } from '@shared/hooks/typed-react-redux-hooks';
 import { FC, useEffect } from 'react';
-import { Outlet } from 'react-router-dom';
 import { push } from 'redux-first-history';
+import { AuthLayout } from '../AuthLayout/AuthLayout';
 interface AuthProviderProps {
     redirect: string;
     passIf: boolean;
@@ -16,5 +16,5 @@ export const AuthProvider: FC<AuthProviderProps> = ({ redirect, passIf }) => {
             dispatch(push(redirect));
         }
     }, [dispatch, redirect, passIf]);
-    return <Outlet />;
+    return <AuthLayout />;
 };
